+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## Unreleased
 
+- (breaking) Fix that `salmon.util/->ex-info` should return an
+  ex-info, not throw it.
 - Fix a bug where stack and change set names could not be refs.
 
 ## v0.22.0 (2025-02-14)